Creating Game-board With Question Page Redirects
Also contains a new background image for the game board and directory organization for the repo. Changed the original CSS file for more specificity when downloading. Also changed Index.php to login.php due to local machine MAMP issues when renaming a file index. Also uploaded a SCRUM template file for Project 2 requirements to keep tabs on our goals for the project.